requirements.txt
# =========================================
# 몽글픽 추천 서비스 의존성
# =========================================

# 웹 프레임워크
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# ORM / 비동기 DB
sqlalchemy[asyncio]>=2.0.0
aiomysql>=0.2.0

# Redis 캐시
redis[hiredis]>=5.0.0

# JWT 인증 (Spring Boot 토큰 검증)
pyjwt>=2.8.0

# 데이터 검증 / 설정
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP 클라이언트
httpx>=0.27.0

# 환경변수 로드
python-dotenv>=1.0.0
